added front facing. working on sign in with apple.

+++ Nitelive/Servicers/UserManager.swift
@@ -26,15 +26,11 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
     }
     @Published var alertItem: AlertItem?
     @Published var profileImage: UIImage? = nil
-    
     @Published var gotUserLocation: Bool = false
     @Published var location: CLLocationCoordinate2D?
     @Published var region = MKCoordinateRegion(center: MapDetails.startingLocation, span: MapDetails.defaultSpan)
-    
-    
-    
+
     var locationManager: CLLocationManager?
-    
   
     override init() {
         super.init()
@@ -43,19 +39,9 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
         }
         
         fetchCurrentUser()
-                
         fetchUserLocation()
-        
-        
-        
-        
+        listenToCheckCurrentUserIsSignedIn()
     }
-
-    /// Fetchs the current user using the app information from Firebase if they are logged in and stores it in UserManager.
-    /// ```
-    /// If the user is logged into Firebase this functions downloads their information and stores
-    /// it in UserManager as well as the FirebaseManager singleton.
-    /// ```
 
     func fetchCurrentUser() {
         guard let uid = FirebaseManager.shared.auth.currentUser?.uid else {
@@ -71,7 +57,6 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
             }
             
             guard let data  = snapshot?.data() else {return}
-                        
             self.currentUser = User(data: data)
             
             if self.currentUser != nil {
@@ -83,20 +68,13 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
         }
     }
     
-    /// Checks if the user is 'Checked In' into any of the clubs and if so downloads that clubs informaiton.
-    /// ```
-    /// The user is able to check in a club based on their proximity of the clubs location and if they
-    ///  are checked in this functions downloads that  clubs information.
-    /// ```
-    
+
     func fetchCurrentClub() {
         
         guard let uid = FirebaseManager.shared.auth.currentUser?.uid else {
             self.errorMessage = "Could not find firebase uid"
             return
         }
-        
-        
         
         FirebaseManager.shared.firestore.collection("users").document(uid).collection(FirebaseConstants.checkedIn).document(FirebaseConstants.checkedInClub)
             .getDocument { snapshot, error in
@@ -106,46 +84,25 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
                 }
                 
                 guard let data = snapshot?.data() else { return }
-                
                 self.currentClub = Club(id: data["id"] as? String ?? "", data: data)
-               
-               
             }
         
     }
     
-    /// Signs the current user out of Firebase
-    /// ```
-    /// This functions signs the current user out of Firebase and sets the current club that ther user is checked in to nil.
-    /// ```
-    
     func handleSignOut() {
         isUserCurrentlyLoggedOut.toggle()
         try? FirebaseManager.shared.auth.signOut()
-        currentClub = nil
-        currentUser = nil
+        checkOutCurrentClub()
     }
     
     
-    /// Checks in the current user into a club.
-    /// ```
-    /// This function first checks if the current user is checked into any other club, and if it is, it checks them out and checks them
-    /// into the current club. It also increments the number of users of user checked into the new club by 1.
-    /// ```
-    
     func checkInCurrentClub(club: Club) {
-      
-       
-        guard let uid = currentUser?.uid else {return}
         
+        guard let uid = currentUser?.uid else {return}
       
         if currentClub != nil {
-            
-        
             if currentClub!.id != club.id {
                 
-               
-             
                 alertItem = AlertContext.checkedOutOfOtherClub
                 
                 self.checkOutCurrentClub()
@@ -155,13 +112,9 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
 
                         if let err = err {
                             print(err)
-                           
                             print("Error setting user data when checking in")
                             return
-
                         }
-                      
-
                         self.currentClub = club
                     }
                 
@@ -175,8 +128,6 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
                             print("error setting club data in user when checking in")
                             return
                         }
-                      
-                    
                     }
    
                 FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(club.id).updateData([FirebaseConstants.checkedIN : FieldValue.increment(Int64(1))]) { err in
@@ -188,23 +139,16 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
                     }
                 }
             }
-               
             
         } else {
-           
             
             FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(club.id).collection(FirebaseConstants.checkedInUsers)
                 .document(uid).setData(currentUser!.asDictionary) { err in
 
                     if let err = err {
                         print(err)
-                       
-
                         return
-
                     }
-                  
-
                     self.currentClub = club
                 }
             
@@ -214,15 +158,10 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
 
                     if let err = err {
                         print(err)
-                      
-
                         return
                     }
-                  
-                
                 }
-            
-            
+        
             FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(club.id).updateData([FirebaseConstants.checkedIN : FieldValue.increment(Int64(1))])
         }
 
@@ -239,16 +178,23 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
     
     func checkOutCurrentClub() {
         print("Checking out....")
-        print(currentClub)
-        print(currentUser)
-        FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(currentClub!.id).collection(FirebaseConstants.checkedInUsers).document(currentUser!.uid).delete()
+
+        if currentUser != nil {
+           
+            FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(currentClub!.id).collection(FirebaseConstants.checkedInUsers).document(currentUser!.uid).delete()
+            
+            
+            FirebaseManager.shared.firestore.collection(FirebaseConstants.users).document(currentUser!.uid).collection(FirebaseConstants.checkedIn).document(FirebaseConstants.checkedInClub).delete()
+            
+            FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(currentClub!.id).updateData([FirebaseConstants.checkedIN : FieldValue.increment(Int64(-1))])
+            
+            self.currentClub = nil
+            self.currentUser = nil
+
+            
+        }
         
-        
-        FirebaseManager.shared.firestore.collection(FirebaseConstants.users).document(currentUser!.uid).collection(FirebaseConstants.checkedIn).document(FirebaseConstants.checkedInClub).delete()
-        
-        FirebaseManager.shared.firestore.collection(FirebaseConstants.locations).document(currentClub!.id).updateData([FirebaseConstants.checkedIN : FieldValue.increment(Int64(-1))])
-        
-        self.currentClub = nil
+       
     }
     
     /// Uploads video to user's current club.
@@ -270,13 +216,6 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
         let storageRef = FirebaseManager.shared.storage.reference(withPath: "\(FirebaseConstants.locationVideos)/\(clubId)/\(videoId)")
         
         storageRef.putFile(from: videoURL, metadata: nil) { metadata, error in
-//          guard let metadata = metadata else {
-//            // Uh-oh, an error occurred!
-//            return
-//          }
-          // Metadata contains file metadata such as size, content-type.
-//          let size = metadata.size
-          // You can also access to download URL after upload.
             
             if let error = error {
                 handler(.failure(error))
@@ -300,41 +239,29 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
               
               handler(.success(true))
               
-          }
-            
-            
+            }
         }
-        
         
     }
     
     
     func fetchUserLocation() {
         checkIfLocationServicesIsEnabled()
-        
     }
-    
    
-    
     func createNotificationForNextEntryToClub(clubId: String, clubLocation: CLLocation) {
         NotificationManager.instance.scheduleLocationNotification(clubId: clubId, clubLocation: clubLocation)
     }
-    
-    
+
     func checkIfLocationServicesIsEnabled() {
         if CLLocationManager.locationServicesEnabled() {
             locationManager = CLLocationManager()
             locationManager!.delegate = self
             locationManager?.desiredAccuracy = kCLLocationAccuracyBest
-            
         } else {
             print("show alert asking them to turn it on")
         }
-        
-        
-        
     }
-    
     
     private func checkLocationAuthorization() {
         guard let locationManager = locationManager else {
@@ -360,13 +287,10 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
                    gotUserLocation = true
                
             }
-             
-          
+            
         @unknown default:
             break
         }
-        
-        
 
     }
     
@@ -392,7 +316,6 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
         
         let logoStorage = FirebaseManager.shared.storage.reference().child("profileImages/\(uid)/\(uid)")
 
-        
         logoStorage.getData(maxSize: 1 * 1024 * 1024) { data, error in
           if let error = error {
             print(error.localizedDescription)
@@ -406,5 +329,16 @@ class UserManager: NSObject, ObservableObject, CLLocationManagerDelegate {
 
         }
         
+    }
+    
+    func listenToCheckCurrentUserIsSignedIn() {
+        let _ = FirebaseManager.shared.auth.addStateDidChangeListener { auth, user in
+            
+            if user == nil {
+                self.checkOutCurrentClub()
+            }
+            
+            
+          }
     }
 }
